Updates to toolchain
Now almost all rt-libraries build. The libc is the worst yet due to
runtime we also need to port and libc has dependencies.

+++ librt/libc/stdio/sscanf.c
@@ -149,7 +149,7 @@ int vsscanf(const char *inp, char const *fmt0, va_list ap)
 		if (c == 0)
 			return (nassigned);
 		if (isspace(c)) {
-			while (inr > 0 && isspace(*inp))
+			while (inr > 0 && isspace((int)(*inp)))
 				nread++, inr--, inp++;
 			continue;
 		}
@@ -284,7 +284,7 @@ int vsscanf(const char *inp, char const *fmt0, va_list ap)
 		* that suppress this.
 		*/
 		if ((flags & NOSKIP) == 0) {
-			while (isspace(*inp)) {
+			while (isspace((int)(*inp))) {
 				nread++;
 				if (--inr > 0)
 					inp++;
@@ -386,7 +386,7 @@ int vsscanf(const char *inp, char const *fmt0, va_list ap)
 				width = (size_t)~0;
 			if (flags & SUPPRESS) {
 				n = 0;
-				while (!isspace(*inp)) {
+				while (!isspace((int)(*inp))) {
 					n++, inr--, inp++;
 					if (--width == 0)
 						break;
@@ -397,7 +397,7 @@ int vsscanf(const char *inp, char const *fmt0, va_list ap)
 			}
 			else {
 				p0 = p = va_arg(ap, char *);
-				while (!isspace(*inp)) {
+				while (!isspace((int)(*inp))) {
 					inr--;
 					*p++ = *inp++;
 					if (--width == 0)
